chore: Update dev scripts

+++ scripts/prepare-dev-win.sh
@@ -1,18 +1,23 @@
 #!/usr/bin/env sh
 
-echo "Cleanup old venv"
-rm -rf env
 
-echo "Setup new venv"
-python -m venv env
+if [ "$1" == "--clean" ]; then
+  echo "Cleanup old venv"
+  rm -rf env
+
+  echo "Setup new venv"
+  python3 -m venv env
+fi
+
 . env/Scripts/activate
 
-echo "Prepare venv"
+echo "Upgrade base tools"
 python -m pip install -U pip setuptools wheel
 
-echo "Install packages:"
+echo ""
+echo "Installing packages"
 for PKG in $(ls packages); do
   echo "- $PKG"
   pip install -r packages/$PKG/requirements.txt
-  pip install -e packages/$PKG
+  pip install packages/$PKG
 done
